Remove the concept of the "current" service period
It's come to light that getting a current service period for today() by
looking for an exception or the first active on today() is incorrect. This
is b/c exceptions can be additive or reductive of some subset of the routes
There's no guarantee that IF there is an exception today(), it will apply
to routes we're interested in.

The quick solution is to implement ServicePeriod.in_service? so that we
can filter routing results based on it. This seems to solve two cases:

a) additive exceptions: trips ADDED as an exception will have the ServicePeriod
   of the exception, therefore we can just check that ServicePeriod i/t filter
b) normal service on a day w/a narrow exception that doesn't apply to all routes

Probably doesn't solve reduced exceptions (holidays)